requirements.txt
black
flake8
isort
loguru
mkdocs
pip
python-dotenv
tqdm
-e .

huggingface_hub
jupyter
click
pandas
matplotlib
scanpy[leiden]
seaborn
timm
dask
tables
mlflow